# docs/openapi.yaml
openapi: 3.0.3
info:
  title: YUANDI ERP API
  description: |
    YUANDI Collection Management System API
    
    해외 구매대행 비즈니스를 위한 통합 ERP 시스템 API입니다.
    주문 관리, 재고 관리, 배송 추적, 현금장부 등의 기능을 제공합니다.
  version: 2.0.0
  contact:
    name: YUANDI Support
    email: api-support@yuandi.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://yuandi-erp.vercel.app/api
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /shipments:
    get:
      tags:
        - Shipments
      summary: List shipments
      description: 배송 목록을 조회합니다
      operationId: getShipments
      parameters:
        - name: status
          in: query
          description: 배송 상태
          schema:
            type: string
            enum: [PREPARING, SHIPPED, DELIVERED, RETURNED]
        - name: courier
          in: query
          description: 택배사 코드
          schema:
            type: string
        - name: tracking_no
          in: query
          description: 운송장 번호
          schema:
            type: string
        - name: date_from
          in: query
          description: 시작 날짜
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          description: 종료 날짜
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Shipment'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
          
    post:
      tags:
        - Shipments
      summary: Create shipment
      description: 새 배송을 생성합니다
      operationId: createShipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShipmentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            total_pages:
              type: integer
...
    Shipment:
      type: object
      properties:
        id:
          type: string
        order_id:
          type: string
        order_no:
          type: string
        tracking_no:
          type: string
        courier:
          type: string
        courier_name:
          type: string
        status:
          type: string
        recipient:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
            address:
              type: string
        shipped_at:
          type: string
          format: date-time
        delivered_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    CreateShipmentRequest:
      type: object
      required:
        - order_id
        - tracking_no
        - courier
      properties:
        order_id:
          type: string
        tracking_no:
          type: string
        courier:
          type: string
        weight_kg:
          type: number
        shipping_cost:
          type: number
        notes:
          type: string
...